Fix uninitialized warnings when modules were not successfully loaded
If a module exsits but can not be loaded(e.g. because of compilation
errors), perl will add the module to %INC anyway, but the corresponding
value is undef instead of the file path.

This commit is to get around uninitialized warnings caused by above
undef values, on module cache update.